requirements.txt
click
Pillow
PyExifTool